summer'25/week2/quint/config/settings.py
import os
from dotenv import load_dotenv
load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY", "mysecret")
GOOGLE_CLIENT_ID = os.getenv("GOOGLE_CLIENT_ID")
GOOGLE_CLIENT_SECRET = os.getenv("GOOGLE_CLIENT_SECRET")
WHATSAPP_TOKEN = os.getenv("WHATSAPP_TOKEN")
LINKEDIN_ACCESS_TOKEN = os.getenv("LINKEDIN_ACCESS_TOKEN")
LOCAL_CONTACTS_PATH = os.getenv("LOCAL_CONTACTS_PATH", "contacts.csv")
