summer'25/week2/quint/requirements.txt
streamlit
agno 
langchain
chromadb
openai
fastapi
uvicorn
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
python-jose[cryptography]
sqlalchemy
passlib[bcrypt]
python-multipart
streamlit
requests
python-dotenv
sqlalchemy
google-auth
google-auth-oauthlib
google-api-python-client
openai
langchain
faiss-cpu
pandas
groq
